+++ DragonBurn/DragonBurn.vcxproj
@@ -236,7 +236,6 @@
     <ClCompile Include="Features\Misc.cpp" />
     <ClCompile Include="Features\Radar.cpp" />
     <ClCompile Include="Features\RCS.cpp" />
-    <ClCompile Include="Features\SpectatorList.cpp" />
     <ClCompile Include="Features\TriggerBot.cpp" />
     <ClCompile Include="Game\Bone.cpp" />
     <ClCompile Include="Game\Entity.cpp" />
