+++ DragonBurn/DragonBurn.vcxproj.filters
@@ -247,9 +247,6 @@
     <ClCompile Include="Features\Radar.cpp">
       <Filter>Features\Visual</Filter>
     </ClCompile>
-    <ClCompile Include="Features\SpectatorList.cpp">
-      <Filter>Features\Misc</Filter>
-    </ClCompile>
     <ClCompile Include="Features\TriggerBot.cpp">
       <Filter>Features\Legit</Filter>
     </ClCompile>
